news_project/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
